Create the daily or monthly COVID-19 cases table of a country or a province

This function is used to create the table to display the daily or monthly COVID-19 cases of a inputted country or province. At the same time, the cumulative number will also be displayed in each table. This function refactors one part of the server-side logic of the Shiny app, which is the table of displaying the daily or monthly COVID-19 cases of a country or a province.

create_cases_table(country_select, province_select = "", cases_frequency)

Arguments

country_select
The name of a country as a character vector. Currently, there are three countries available, which are "Japan", "US", and "Australia".

province_select
The name of a province as a character vector. Currently, only Australia has the provincial data. The available provinces are "Australian Capital Territory", "New South Wales", "Northern Territory", "Queensland", "South Australia", "Tasmania", "Victoria", and "Western Australia".

cases_frequency
A character vector indicates the frequency of cases in the table. The available choices are "monthly" or "daily".

Value

A table that shows the daily or monthly COVID-19 cases and the cumulative cases of a country or a province.

Note

Only Australia has the available provincial data, please do not input any character string into the argument of province_select when choosing "Japan" or "US".

Examples

create_cases_table(country_select = "Japan", cases_frequency = "monthly")
create_cases_table(country_select = "Australia", province_select = "Victoria", cases_frequency = "monthly")
create_cases_table(country_select = "US", cases_frequency = "daily")