Select the daily COVID-19 cases of a country within a month

This function is used to find the three different daily COVID-19 cases of the selected country in a specific month. This function refactors one part of the user interface-side (UI) logic of the Shiny app, which is the selection bar of selecting the country and a specific month.

select_country_month(country_select, month_select)

Arguments

country_select
The name of a country as a character vector. Currently, there are three countries available, which are "Japan", "US", and "Australia".

month_select
The number of a month as an integer vector. The available number of month is between 1 and 10.

Format

This function will return a tibble with 7 columns.

country

Name of the country.

year, month, day

The number of the year, month and day.

confirmed

The number of daily confirmed cases of COVID-19.

death

The number of daily mortality of COVID-19.

recovered

The number of daily recovered cases of COVID-19.

Value

A tibble that contains the information of country, year, month, day, and daily COVID-19 cases.

Examples

select_country_month(country_select = "Japan", month_select = 1)

#> # A tibble: 10 x 7
#> # Groups:   country [1]
#>    country  year month   day confirmed death recovered
#>    <chr>   <dbl> <dbl> <int>     <int> <int>     <int>
#>  1 Japan    2020     1    22         2     0         0
#>  2 Japan    2020     1    23         0     0         0
#>  3 Japan    2020     1    24         0     0         0
#>  4 Japan    2020     1    25         0     0         0
#>  5 Japan    2020     1    26         2     0         1
#>  6 Japan    2020     1    27         0     0         0
#>  7 Japan    2020     1    28         3     0         0
#>  8 Japan    2020     1    29         0     0         0
#>  9 Japan    2020     1    30         4     0         0
#> 10 Japan    2020     1    31         4     0         0

select_country_month(country_select = "US", month_select = 3)

#> # A tibble: 31 x 7
#> # Groups:   country [1]
#>    country  year month   day confirmed death recovered
#>    <chr>   <dbl> <dbl> <int>     <int> <int>     <int>
#>  1 US       2020     3     1         6     0         0
#>  2 US       2020     3     2        23     5         0
#>  3 US       2020     3     3        20     1         0
#>  4 US       2020     3     4        31     4         0
#>  5 US       2020     3     5        70     1         0
#>  6 US       2020     3     6        48     2         0
#>  7 US       2020     3     7       115     3         0
#>  8 US       2020     3     8       114     4         0
#>  9 US       2020     3     9        68     1         0
#> 10 US       2020     3    10       192     6         1
#> # … with 21 more rows

select_country_month(country_select = "Australia", month_select = 5)

#> # A tibble: 31 x 7
#> # Groups:   country [1]
#>    country    year month   day confirmed death recovered
#>    <chr>     <dbl> <dbl> <int>     <int> <int>     <int>
#>  1 Australia  2020     5     1        12     0        33
#>  2 Australia  2020     5     2        21     1        39
#>  3 Australia  2020     5     3        23     1        35
#>  4 Australia  2020     5     4        25     1        38
#>  5 Australia  2020     5     5        28     1        88
#>  6 Australia  2020     5     6        19     0        56
#>  7 Australia  2020     5     7        19     0        47
#>  8 Australia  2020     5     8         5     0        44
#>  9 Australia  2020     5     9        21     0        19
#> 10 Australia  2020     5    10         9     0        26
#> # … with 21 more rows