Select the monthly COVID-19 cases of a country within a range of months

This function is used to find the three different monthly COVID-19 cases of the selected country within the specific range of months. This function refactors one part of the user interface-side (UI) logic of the Shiny app, which is the selection bar of selecting the country and the range of months.

select_country_months(country_select, start_month, end_month)

Arguments

country_select
The name of a country as a character vector. Currently, there are three countries available, which are "Japan", "US", and "Australia".

start_month
The number of a month that represents the start of the range of months. It should be an integer vector, and the available months are between 1 and 10.

end_month
The number of a month that represents the end of the range of months. It should be an integer vector, and the available months are between 1 and 10.

Format

This function will return a tibble with 8 columns.

country

Name of country.

month

The range of months. Each row represent a month.

confirmed

Monthly confirmed cases of COVID-19 .

death

Monthly mortality of COVID-19.

recovered

Monthly recovered cases of COVID-19.

cumulative_confirmed

Cumulative confirmed cases of current month, which calculated from January.

cumulative_death

Cumulative mortality of current month, which calculated from January.

cumulative_recovered

Cumulative recovered cases of current month, which calculated from January.

Value

A tibble that contains the information of country, month, monthly cases, and the cumulative cases of COVID-19.

Examples

select_country_months(country_select = "Japan", start_month = 1, end_month = 10)

#> # A tibble: 7 x 8
#> # Groups:   country [1]
#>   country month confirmed death recovered cumulative_conf… cumulative_death
#>   <chr>   <dbl>     <int> <int>     <int>            <int>            <int>
#> 1 Japan       1        15     0         1               15                0
#> 2 Japan       2       230     6        31              245                6
#> 3 Japan       3      2010    61       392             2255               67
#> 4 Japan       4     12029   414      2036            14284              481
#> 5 Japan       5      2467   417     11882            16751              898
#> 6 Japan       6      1864    74      2116            18615              972
#> 7 Japan       7     17619    36      8316            36234             1008
#> # … with 1 more variable: cumulative_recovered <int>

select_country_months(country_select = "US", start_month = 3, end_month = 9)

#> # A tibble: 5 x 8
#> # Groups:   country [1]
#>   country month confirmed death recovered cumulative_conf… cumulative_death
#>   <chr>   <dbl>     <int> <int>     <int>            <int>            <int>
#> 1 US          3    188700  5604      7017           188724             5605
#> 2 US          4    883943 58651    146923          1072667            64256
#> 3 US          5    726457 41108    290811          1799124           105364
#> 4 US          6    837290 22068    275873          2636414           127432
#> 5 US          7   1925624 25882    717529          4562038           153314
#> # … with 1 more variable: cumulative_recovered <int>

select_country_months(country_select = "Australia", start_month = 5, end_month = 6)

#> # A tibble: 2 x 8
#> # Groups:   country [1]
#>   country month confirmed death recovered cumulative_conf… cumulative_death
#>   <chr>   <dbl>     <int> <int>     <int>            <int>            <int>
#> 1 Austra…     5       436    10       876             7202              103
#> 2 Austra…     6       718     1       422             7920              104
#> # … with 1 more variable: cumulative_recovered <int>